+++ ecd-20200930/alibabacloud_ecd20200930/models/_describe_desktops_response_body.py
@@ -115,6 +115,8 @@ class DescribeDesktopsResponseBodyDesktops(DaraModel):
         downgraded_times: int = None,
         end_user_ids: List[str] = None,
         entra_domain_name: str = None,
+        env_id: str = None,
+        env_type: str = None,
         expired_time: str = None,
         fota_update: main_models.DescribeDesktopsResponseBodyDesktopsFotaUpdate = None,
         gpu_category: int = None,
@@ -235,6 +237,8 @@ class DescribeDesktopsResponseBodyDesktops(DaraModel):
         # The end user IDs.
         self.end_user_ids = end_user_ids
         self.entra_domain_name = entra_domain_name
+        self.env_id = env_id
+        self.env_type = env_type
         # The time when a subscription cloud computer expired.
         self.expired_time = expired_time
         # The information about the image version of the cloud computer.
@@ -477,6 +481,12 @@ class DescribeDesktopsResponseBodyDesktops(DaraModel):
         if self.entra_domain_name is not None:
             result['EntraDomainName'] = self.entra_domain_name
 
+        if self.env_id is not None:
+            result['EnvId'] = self.env_id
+
+        if self.env_type is not None:
+            result['EnvType'] = self.env_type
+
         if self.expired_time is not None:
             result['ExpiredTime'] = self.expired_time
 
@@ -696,6 +706,12 @@ class DescribeDesktopsResponseBodyDesktops(DaraModel):
 
         if m.get('EntraDomainName') is not None:
             self.entra_domain_name = m.get('EntraDomainName')
+
+        if m.get('EnvId') is not None:
+            self.env_id = m.get('EnvId')
+
+        if m.get('EnvType') is not None:
+            self.env_type = m.get('EnvType')
 
         if m.get('ExpiredTime') is not None:
             self.expired_time = m.get('ExpiredTime')